Исправляем бэги

+++ utils/misc/throttling.py
@@ -1,4 +1,4 @@
-def rate_limit(limit: int, key=None):
+def rate_limit(limit: float, key=None):
     """
     Decorator for configuring rate limit and key in different functions.
